Add stateful menu using :checked pseudo selector
This (ab)uses a checkbox to hold ephemeral state combined with sibling
selectors like the default pane.

It features a nice dimmed background that transitions along with the
menu. However, that may look easier then one thinks. We need to cover
the entire page with a label for the menu dismissal to work. But we
can't use that label element for the background because it needs to
resize during the transition. First it should only cover the menu
button, then the entire area. This results in janky animation when
dissmising the manu.

Fortunately we can use the new pointer-events property to place an
invisible, non-interactive pseudo element on top of the entire page and
style that as appropriate.